cleanup command now also deletes redundant meetup threads
