+++ src/command/MeetupCleanUpCommand.ts
@@ -6,6 +6,7 @@ import { tCommand } from "../i18n";
 import { assertMeetupCreateChannelUsed } from "../permission/assertMeetupCreateChannelUsed";
 import { assertUserHasMeetupConfigRole } from "../permission/assertUserHasMeetupConfigRole";
 import { cleanupMeetupData } from "../util/cleanupMeetupData";
+import { deleteRedundantMeetupThreads } from "../util/deleteRedundantMeetupThreads";
 import { postSuccess } from "../util/postEmbeds";
 import { AbstractCommand } from "./AbstractCommand";
 
@@ -30,6 +31,7 @@ export class MeetupCleanUpCommand extends AbstractCommand {
         await interaction.deferReply({ flags: MessageFlags.Ephemeral });
 
         await cleanupMeetupData();
+        await deleteRedundantMeetupThreads();
 
         //create success embed
         await postSuccess(interaction, tCommand("meetupCleanup.success"));
